test volume replacement with flaky ample
The goal of these changes is to excercise the conditional mutation code
used to replace volumes in tablets.

Refactored some code in VolumeIT into VolumeBaseIT so that VolumeBaseIT
could be extended by VolumeFlakyAmpleIT that injects manager and
tservers that use FlakyAmple. Ran into problems during this change.
First problem was that manager and tserver processes were dying because
of OOME. Using heap dumps found that TestAmple objects were continually
being created and each TestAmple object created a hadoop configuration
object. Eventually memory would fill up with these configuration
objects. Do not fully understand why the hadoop config object never
went away because it seemed like the TestAmple objects should have been
transient. This problem may be worth further investigation.  In this
commit only a single TestAmple object per ServerContext is created to
solve the problem.

The second problem encountered was that VolumeIT was created 2 tables
with 100 tablets each.  When using FlakyAmple all tablet operations take
longer and the test ended up timing out because of this.  To solve this
problem, lowered the number of tablet and increased the number of
compactors.
