+++ test/src/main/java/org/apache/accumulo/test/ample/FlakyAmpleServerContext.java
@@ -19,11 +19,14 @@
 package org.apache.accumulo.test.ample;
 
 import java.util.Map;
+import java.util.function.Supplier;
 
 import org.apache.accumulo.core.conf.SiteConfiguration;
 import org.apache.accumulo.core.metadata.schema.Ample;
 import org.apache.accumulo.server.ServerContext;
 import org.apache.accumulo.test.ample.metadata.TestAmple;
+
+import com.google.common.base.Suppliers;
 
 /**
  * A goal of this class is to exercise the lambdas passed to
@@ -33,13 +36,22 @@ import org.apache.accumulo.test.ample.metadata.TestAmple;
  */
 public class FlakyAmpleServerContext extends ServerContext {
 
+  private final Supplier<Ample> ampleSupplier;
+
   public FlakyAmpleServerContext(SiteConfiguration siteConfig) {
     super(siteConfig);
+    // Each instance of TestAmple created will create a new Haddop configuration object. These
+    // seemed to hang around and cause OOME and process death. Did not track down why they were
+    // hanging around, but decided to avoid creating a new instance of TestAmple each time Ample is
+    // requested in order to avoid creating those hadoop config objects.
+    ampleSupplier = Suppliers.memoize(() -> TestAmple.create(
+        this, Map.of(Ample.DataLevel.USER, Ample.DataLevel.USER.metaTable(),
+            Ample.DataLevel.METADATA, Ample.DataLevel.METADATA.metaTable()),
+        FlakyInterceptor::new));
   }
 
   @Override
   public Ample getAmple() {
-    return TestAmple.create(this, Map.of(Ample.DataLevel.USER, Ample.DataLevel.USER.metaTable(),
-        Ample.DataLevel.METADATA, Ample.DataLevel.METADATA.metaTable()), FlakyInterceptor::new);
+    return ampleSupplier.get();
   }
 }
